pyproject.toml
[tool.poetry]
name = "vitap-vtop-api"
version = "0.1.0"
description = "A FastAPI wrapper for vitap-vtop-client."
authors = ["Udhay Adithya udhayxd@gmail.com"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13.3"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.34.2"
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
starlette = "^0.46.2"
gunicorn = "^23.0.0"
vitap-vtop-client = { git = "https://github.com/Udhay-Adithya/vitap-vtop-client.git", rev = "main" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
